+++ doc/changelog.rst
@@ -24,11 +24,19 @@ Addons
 Front
 ~~~~~
 
+- Fix macro name in Single Page Checkout
+- Add Saved Carts to Dashboard
+- Add Order History to Dashboard
+- Add Customer Information to Dashboard
+- Add Dashboard for customers
+
 Xtheme
 ~~~~~~
 
 Classic Gray Theme
 ~~~~~~~~~~~~~~~~~~
+
+- Fix issue with footer padding
 
 Simple Supplier
 ~~~~~~~~~~~~~~~
@@ -60,8 +68,12 @@ Importer
 Regions
 ~~~~~~~
 
+- Make backend more modular to allow more specific resource distribution
+
 General/miscellaneous
 ~~~~~~~~~~~~~~~~~~~~~
+
+- Personal Order history: URL has now been changed from ``/orders`` to ``/order-history``
 
 SHUUP 0.5.0
 -----------